Move initial Google Tasks code into own class
- Separate Get Lists into own method
